impressum-scraper/requirements.txt
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pyyaml>=6.0.0
python-dotenv>=1.0.0
lxml>=4.9.0
urllib3>=2.0.0
